+++ server/src/middleware/paramSchema.js
@@ -1,0 +1,11 @@
+import Joi from 'joi';
+
+const params = Joi.object({
+  id: Joi.string().guid({
+    version: [
+      'uuidv4'
+    ]
+  })
+});
+
+export default params;
